Add ajax for creating comments on questions and answers

+++ app/controllers/answer.rb
@@ -16,27 +16,50 @@ get '/answers/:id/vote' do
   answer = Answer.find(params[:id])
   redirect "/questions/#{answer.question.id}" if !logged_in?
   if answer.votes.find_by(user_id: current_user.id) != nil
-    redirect "/questions/#{answer.question.id}"
+    if request.xhr?
+      answer.total_votes.to_s
+    else
+      redirect "/questions/#{answer.question.id}"
+    end
   else
     new_vote = Vote.new(voter: current_user)
     answer.votes << new_vote
+   if request.xhr?
+      answer.total_votes.to_s
+    else
+      redirect "/questions/#{answer.question.id}"
+    end
+  end
+
+  if request.xhr?
+    answer.total_votes.to_s
+  else
     redirect "/questions/#{answer.question.id}"
   end
 end
 
 get '/answers/:id/delete-vote' do
-  question = Answer.find(params[:id])
+  answer = Answer.find(params[:id])
   redirect "/answers/#{answer.id}" if !logged_in?
   vote = answer.votes.find_by(user_id: current_user.id)
-  vote.destroy
-  redirect "/questions/#{answer.question.id}"
+  vote.destroy if vote
+  if request.xhr?
+    answer.total_votes.to_s
+  else
+    redirect "/questions/#{answer.question.id}"
+  end
 end
 
 post '/answers/:id/comments' do
   answer = Answer.find(params[:id])
-  answer.comments.create(body: params[:comment][:body], commentor: current_user)
+  new_comment = answer.comments.create(body: params[:comment][:body], commentor: current_user)
 
-  redirect "/questions/#{answer.question.id}"
+  if request.xhr?
+    content_type :json
+    {id: answer.id, body: new_comment.body, username: new_comment.commentor.username}.to_json
+  else
+    redirect "/questions/#{answer.question.id}"
+  end
 
 end
